test: e2e ticket fix
